♻️ Refactor the usages of the policy repository interface to use new design.
