+++ gearmeshing_ai/agent_core/policy/provider.py
@@ -4,7 +4,7 @@ import logging
 from dataclasses import dataclass
 from typing import Dict, Protocol
 
-from ..repos.interfaces import PolicyRepository
+from gearmeshing_ai.core.database.repositories.policies import PolicyRepository
 from gearmeshing_ai.core.models.domain import AgentRun
 from gearmeshing_ai.core.models.domain.policy import PolicyConfig
