+++ test/integration_test/agent_core/test_policy_persistence_integration.py
@@ -18,7 +18,7 @@ from gearmeshing_ai.agent_core.policy.provider import (
     StaticPolicyProvider,
     async_get_policy_config,
 )
-from gearmeshing_ai.agent_core.repos.interfaces import PolicyRepository
+from gearmeshing_ai.core.database.repositories.policies import PolicyRepository
 from gearmeshing_ai.core.models.domain import (
     AgentRun,
     AgentRunStatus,
